pyproject.toml
[tool.poetry]
name = "dafiti_challenge"
version = "0.1.0"
description = ""
authors = ["Mateus Lisboa <mateuslnobre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
ipython = "^8.1.1"
django-extensions = "^3.1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"